# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  routes.AddModDTO:
    properties:
      userId:
        type: string
    required:
    - userId
    type: object
...
  routes.CreateGroupDTO:
    properties:
      bannerUrl:
        type: string
      description:
        type: string
      groupName:
        type: string
    required:
    - groupName
    type: object
...
  routes.GroupDTO:
    properties:
      bannerUrl:
        type: string
      description:
        type: string
      groupName:
        type: string
      id:
        type: string
      registeredAt:
        type: string
    type: object
  routes.GroupDetailDTO:
    properties:
      bannerUrl:
        type: string
      description:
        type: string
      groupName:
        type: string
      id:
        type: string
      moderators:
        items:
          $ref: '#/definitions/routes.UserProfile'
        type: array
      registeredAt:
        type: string
      users:
        items:
          $ref: '#/definitions/routes.UserProfile'
        type: array
    type: object
...
  routes.UpdateGroupDTO:
    properties:
      bannerUrl:
        type: string
      description:
        type: string
    type: object
...
  routes.UserProfile:
    properties:
      bannerUrl:
        type: string
      commentReputation:
        type: integer
      email:
        type: string
      id:
        type: string
      nickname:
        type: string
      postReputation:
        type: integer
      registeredAt:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a Reddit-like REST API.
  title: Chirp API
  version: "1.0"
paths:
...
  /groups:
    get:
      description: Получает список всех групп
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/routes.GroupDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить список групп
      tags:
      - groups
    post:
      consumes:
      - application/json
      description: Создаёт новую группу
      parameters:
      - description: Данные для группы
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/routes.CreateGroupDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/routes.GroupDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Создать группу
      tags:
      - groups
  /groups/{groupId}/moderators:
    post:
      consumes:
      - application/json
      description: Добавляет пользователя в список модераторов группы
      parameters:
      - description: ID группы
        in: path
        name: groupId
        required: true
        type: string
      - description: Данные модератора
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/routes.AddModDTO'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Добавить модератора в группу
      tags:
      - moderation
  /groups/{groupId}/moderators/{userId}:
    delete:
      description: Удаляет пользователя из списка модераторов группы
      parameters:
      - description: ID группы
        in: path
        name: groupId
        required: true
        type: string
      - description: ID пользователя
        in: path
        name: userId
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удалить модератора из группы
      tags:
      - moderation
...
  /groups/{id}:
    delete:
      description: Удаляет группу
      parameters:
      - description: ID группы
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удалить группу
      tags:
      - groups
    get:
      description: Получает подробную информацию о группе
      parameters:
      - description: ID группы
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.GroupDetailDTO'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить детали группы
      tags:
      - groups
    put:
      consumes:
      - application/json
      description: Обновляет данные группы
      parameters:
      - description: ID группы
        in: path
        name: id
        required: true
        type: string
      - description: Данные для обновления
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/routes.UpdateGroupDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/routes.GroupDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновить группу
      tags:
      - groups
...
swagger: "2.0"